refactor : 코드래빗 리뷰 반영

+++ src/main/java/com/gitggal/clothesplz/dto/user/UserDtoCursorResponse.java
@@ -8,7 +8,7 @@ public record UserDtoCursorResponse(
     String nextCursor,
     UUID nextIdAfter,
     boolean hasNext,
-    int totalCount,
+    long totalCount,
     String sortBy,
     String sortDirection
 ) {
